requirements.txt
asgiref==3.8.1
Django==5.2.3
django-bootstrap5==25.1
django-decouple==2.1
pillow==11.2.1
sqlparse==0.5.3
joblib==1.5.1
numpy==1.24.3
pandas==2.3.0
scikit-surprise==1.1.4
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
python-dateutil==2.9.0.post0
pytz==2025.2
